+++ Pandora/Paradigm/Primary/Transformer/Backwards.hs
@@ -7,7 +7,7 @@ import Pandora.Pattern.Functor.Extractable (Extractable (extract))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)))
 import Pandora.Pattern.Functor.Traversable (Traversable ((->>)))
-import Pandora.Pattern.Functor.Distributive (Distributive ((>>-)))
+import Pandora.Pattern.Functor.Distributive (Distributive_ ((--<<-)))
 import Pandora.Pattern.Transformer.Liftable (Liftable (lift))
 import Pandora.Pattern.Transformer.Lowerable (Lowerable (lower))
 import Pandora.Pattern.Transformer.Hoistable (Hoistable ((/|\)))
@@ -34,8 +34,8 @@ instance Applicative t => Applicative (Backwards t) where
 instance Traversable t => Traversable (Backwards t) where
 	Backwards x ->> f = Backwards <$> x ->> f
 
-instance Distributive t => Distributive (Backwards t) where
-	x >>- f = Backwards $ x >>- run . f
+instance Distributive_ t (->) (->) => Distributive_ (Backwards t) (->) (->) where
+	f --<<- x = Backwards $ run . f --<<- x
 
 instance Contravariant t => Contravariant (Backwards t) where
 	f >$< Backwards x = Backwards $ f >$< x