+++ Pandora/Paradigm/Primary/Transformer/Reverse.hs
@@ -9,7 +9,7 @@ import Pandora.Pattern.Functor.Extractable (Extractable (extract))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)))
 import Pandora.Pattern.Functor.Traversable (Traversable ((->>)))
-import Pandora.Pattern.Functor.Distributive (Distributive ((>>-)))
+import Pandora.Pattern.Functor.Distributive (Distributive_ ((--<<-)))
 import Pandora.Pattern.Transformer.Liftable (Liftable (lift))
 import Pandora.Pattern.Transformer.Lowerable (Lowerable (lower))
 import Pandora.Pattern.Transformer.Hoistable (Hoistable ((/|\)))
@@ -36,8 +36,8 @@ instance Applicative t => Applicative (Reverse t) where
 instance Traversable t => Traversable (Reverse t) where
 	Reverse x ->> f = Reverse <$> run (x ->> Backwards . f)
 
-instance Distributive t => Distributive (Reverse t) where
-	x >>- f = Reverse $ x >>- run . f
+instance Distributive_ t (->) (->) => Distributive_ (Reverse t) (->) (->) where
+	f --<<- x = Reverse $ run . f --<<- x
 
 instance Contravariant t => Contravariant (Reverse t) where
 	f >$< Reverse x = Reverse # f >$< x